+++ pyproject.toml
@@ -84,8 +84,7 @@ dependencies = [
   "mlflow",
 
   # Elasticsearch
-  "elasticsearch>=7.7,<7.11",
-  "elastic-apm",
+  "elasticsearch>=7.7,<8",
 
   # context matching
   "rapidfuzz>=2.0.15,<2.8.0",   # FIXME https://github.com/deepset-ai/haystack/pull/3199
